Separate node-follower workspace
- Add node_follower directory
- Common behavior only exists in node_follower/src/lib.rs
- Update mobile_rewards to use node_follower
- Update poc_iot_verifier to use node_follower
