+++ mobile_rewards/src/traits/mod.rs
@@ -1,9 +1,11 @@
 pub mod b64;
+pub mod get_reward_period;
 pub mod owner_resolver;
 pub mod txn_hash;
 pub mod txn_sign;
 
 pub use b64::B64;
+pub use get_reward_period::{GetRewardPeriod, RewardPeriod};
 pub use owner_resolver::OwnerResolver;
 pub use txn_hash::TxnHash;
 pub use txn_sign::TxnSign;
