+++ poc_iot_verifier/src/error.rs
@@ -30,6 +30,8 @@ pub enum Error {
     NotFound(String),
     #[error("base64 decode error")]
     Base64DecodeError(#[from] base64::DecodeError),
+    #[error("follower error")]
+    Follower(#[from] node_follower::Error),
 }
 
 #[derive(Error, Debug)]
