+++ poc_iot_verifier/src/lib.rs
@@ -12,7 +12,6 @@ pub mod traits;
 
 pub use error::{Error, Result};
 use sqlx::{postgres::PgPoolOptions, Pool, Postgres};
-use std::io;
 
 pub async fn mk_db_pool(size: u32) -> Result<Pool<Postgres>> {
     let db_connection_str = dotenv::var("DATABASE_URL")?;
@@ -21,18 +20,4 @@ pub async fn mk_db_pool(size: u32) -> Result<Pool<Postgres>> {
         .connect(&db_connection_str)
         .await?;
     Ok(pool)
-}
-
-fn env_var<T>(key: &str, default: T) -> Result<T>
-where
-    T: std::str::FromStr,
-    <T as std::str::FromStr>::Err: std::fmt::Debug,
-{
-    match dotenv::var(key) {
-        Ok(v) => v
-            .parse::<T>()
-            .map_err(|_err| Error::from(io::Error::from(io::ErrorKind::InvalidInput))),
-        Err(dotenv::Error::EnvVar(std::env::VarError::NotPresent)) => Ok(default),
-        Err(err) => Err(Error::from(err)),
-    }
 }
