+++ poc_iot_verifier/src/traits/gateway_info_resolver.rs
@@ -1,0 +1,9 @@
+use crate::Result;
+use async_trait::async_trait;
+use helium_crypto::PublicKey;
+use node_follower::gateway_resp::FollowerGatewayResp;
+
+#[async_trait]
+pub trait GatewayInfoResolver {
+    async fn resolve_gateway_info(&mut self, address: &PublicKey) -> Result<FollowerGatewayResp>;
+}
