+++ poc_iot_verifier/src/traits/mod.rs
@@ -1,2 +1,5 @@
 pub mod b64;
+pub mod gateway_info_resolver;
+
 pub use b64::B64;
+pub use gateway_info_resolver::GatewayInfoResolver;
